Consolidate PublicApiTests code so we don't repeat ourselves
